@@ -21,13 +21,13 @@ def getBranchName = { ->
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "nl.jeroenhd.app.bcbreader"
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 7
         versionName "Elementary " + getBranchName() + "@" + getVersionName() + ' (' + new Date().format("yyyyMMddHHmmss") + ')'
     }
@@ -49,6 +49,7 @@ android {
 }
 
 def dbflow_version = "3.1.1"
+def work_version = "2.0.0"
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
@@ -58,14 +59,15 @@ dependencies {
     implementation "com.github.Raizlabs.DBFlow:dbflow-core:${dbflow_version}"
     implementation "com.github.Raizlabs.DBFlow:dbflow:${dbflow_version}"
 
+    // (Java only)
+    implementation "androidx.work:work-runtime:$work_version"
+
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:palette-v7:27.1.1'
-    implementation 'com.google.code.gson:gson:2.8.2'
-    implementation 'com.android.volley:volley:1.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
-    implementation 'com.evernote:android-job:1.2.5'
-    testImplementation 'org.mockito:mockito-core:2.8.9'
+    implementation 'com.android.support:design:28.0.0'
+    implementation 'com.android.support:palette-v7:28.0.0'
+    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.android.volley:volley:1.1.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    testImplementation 'org.mockito:mockito-core:2.23.0'
     androidTestImplementation 'org.testng:testng:6.9.6'
 }
