@@ -1,12 +1,13 @@
 package nl.jeroenhd.app.bcbreader.adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import nl.jeroenhd.app.bcbreader.R;
 import nl.jeroenhd.app.bcbreader.data.Chapter;
 import nl.jeroenhd.app.bcbreader.views.PageThumbImageView;
@@ -36,15 +37,16 @@ public class PageThumbAdapter extends RecyclerView.Adapter<PageThumbAdapter.View
             this.notifyItemRangeInserted(0, mChapter.getPageCount());
     }
 
+    @NonNull
     @Override
-    public PageThumbAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public PageThumbAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         LayoutInflater inflater = LayoutInflater.from(mContext);
         View inflatedView = inflater.inflate(R.layout.content_page_thumb, parent, false);
         return new PageThumbAdapter.ViewHolder(inflatedView, this.onThumbClickListener);
     }
 
     @Override
-    public void onBindViewHolder(PageThumbAdapter.ViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull PageThumbAdapter.ViewHolder holder, int position) {
         holder.pageText.setText(String.format(mContext.getString(R.string.page_number), position + 1));
         holder.pageThumbImageView.setPage(mChapter.getNumber(), position + 1);
         holder.page = position + 1;
@@ -78,14 +80,9 @@ public class PageThumbAdapter extends RecyclerView.Adapter<PageThumbAdapter.View
         ViewHolder(final View itemView, OnThumbClickListener onThumbClickListener) {
             super(itemView);
 
-            this.pageThumbImageView = (PageThumbImageView) itemView.findViewById(R.id.thumb);
-            this.pageThumbImageView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    onClick.onThumbnailClick(mChapter, page, v);
-                }
-            });
-            this.pageText = (TextView) itemView.findViewById(R.id.text);
+            this.pageThumbImageView = itemView.findViewById(R.id.thumb);
+            this.pageThumbImageView.setOnClickListener(v -> onClick.onThumbnailClick(mChapter, page, v));
+            this.pageText = itemView.findViewById(R.id.text);
             this.onClick = onThumbClickListener;
         }
     }
