@@ -7,12 +7,10 @@ import android.content.Context;
 import android.graphics.Color;
 import android.util.Log;
 
-import com.evernote.android.job.JobManager;
 import com.raizlabs.android.dbflow.config.FlowConfig;
 import com.raizlabs.android.dbflow.config.FlowManager;
 
-import nl.jeroenhd.app.bcbreader.notifications.CheckForUpdateJob;
-import nl.jeroenhd.app.bcbreader.notifications.NotificationJobCreator;
+import nl.jeroenhd.app.bcbreader.notifications.CheckForUpdateWorker;
 import nl.jeroenhd.app.bcbreader.tools.AppCrashStorage;
 import nl.jeroenhd.app.bcbreader.tools.Shortcuts;
 
@@ -46,8 +44,7 @@ public class BCBReaderApplication extends Application {
 
         // Prepare notifications
         this.prepareNotifications();
-        JobManager.create(this).addJobCreator(new NotificationJobCreator());
-        CheckForUpdateJob.schedule(this);
+        CheckForUpdateWorker.schedule(this);
 
         Shortcuts.Update(this);
     }
