@@ -3,12 +3,13 @@ package nl.jeroenhd.app.bcbreader.data;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
 import android.util.Log;
 
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
+
+import androidx.annotation.NonNull;
 
 /**
  * API URL container
@@ -95,7 +96,11 @@ public class API {
             return "@m";
         } else {
             String p = sharedPreferences.getString("reading_quality", "-1");
-            int quality = Integer.parseInt(p);
+            int quality = 0;
+
+            if (p != null) {
+                quality = Integer.parseInt(p);
+            }
 
             switch (quality) {
                 case -1:
@@ -151,21 +156,6 @@ public class API {
     public static String FormatPageUrl(Double chapter, double page, String quality) {
         String ext = isJpegChapter(chapter, quality) ? ".jpg" : ".png";
 
-        /**
-         * Check if we need JPG of PNG files for the chapter pages
-         * Proposed fix: use CDN, leave out extension
-         * Not done because: doesn't work
-         *
-         * 1:31 PM <jeroenhd> Quick question: I was working on my Android app and I noticed some chapters are JPG files and some are PNG files. I thought the JPGs were only for some old chapters but then I saw that 70-88 are also JPG chapters. For now I have hardcoded which chapters are JPG and which are PNG, but I fear the app will break if a future chapter will be a JPG chapter again.
-         * 1:31 PM <jeroenhd> On the old server/CDN leaving out the extension redirected the client to the correct file (so /comics/88/1 would redirect to /comics/88/1.png or /comics/88/1.jpg automatically). Is there any API somewhere that helps determining what extension the images have? I could always code it so it tries the PNG and falls back to the JPG but that requires extra HTTP
-         * 1:31 PM <jeroenhd> requests/network delays/battery life etc.
-         * 1:59 PM <SuitCase> jpg\png should be in the api somehow? i dunno. it’s certainly in the bcb_comics mysql
-         * 1:59 PM <SuitCase> i somehow resolved this for the iphone i thought
-         * 1:59 PM <SuitCase> i thought i wrote a thing for nginx that does the auto resolution where it checks if it’s jpg or png
-         * 2:00 PM <jeroenhd> I use the blasto.enterprises CDN but that spits out a 404 if I leave out the extension
-         * 2:00 PM <SuitCase> dont use that
-         * 2:00 PM <SuitCase> use the site itself, https://www.
-         **/
         return CDNUrl + "comics/" + FormatChapterNumber(chapter) + "/" + "" + ((long) page) + quality + ext;
     }
 
@@ -239,7 +229,7 @@ public class API {
      */
     @Deprecated
     public static String FormatChapterThumbURL(Double chapter) {
-        /***
+        /*
          * CDN returns 404 page instead of empty image for unknown chapters
          * This is the behaviour the app is designed for for now. This might change though,
          * so the code remains here for now.
