@@ -2,7 +2,6 @@ package nl.jeroenhd.app.bcbreader.data;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.Nullable;
 
 import com.google.gson.annotations.Expose;
 import com.raizlabs.android.dbflow.annotation.Column;
@@ -17,6 +16,7 @@ import com.raizlabs.android.dbflow.structure.BaseModel;
 import java.util.ArrayList;
 import java.util.List;
 
+import androidx.annotation.Nullable;
 import nl.jeroenhd.app.bcbreader.data.databases.ChapterDatabase;
 
 /**
